pyproject.toml
[tool.poetry]
name = "twitter_watcher"
version = "0.0.1"
description = ""
authors = ["sasano8 <y-sasahara@ys-method.com>"]
license = "MIT"
readme = "README.md"
# repository = "https://github.com/sasano8/asy"

[tool.poetry.dependencies]
python = "^3.8"
asy = "^0.0.7"
pydantic = "^1.8.2"
tweepy = "^3.10.0"
python-dotenv = "^0.17.1"
line-bot-sdk = "^1.19.0"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
mypy = "^0.812"
pytest = "^6.2.2"
pre-commit = "^2.12.0"
devtools = "^0.6.1"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
